Create reverse workflow analysis.sql
This identifies cases when articles are sent backward in the workflow from copy edit > initial edit. This creative solution for identifying reverse workflow instances is necessary since direction of submissions (forward or backward) are not indicated in the logs tables. In the future, we may work with the dev team to capture this data in the logs tables if there becomes a greater reporting need from editorial. Identifying reverse workflow instances and frequency is important for understanding where there can be improvements can be made in the Mercury workflow.